dashboard: structure rearrangement and polish related code
