+++ sentinel-dashboard/src/main/java/com/taobao/csp/sentinel/dashboard/controller/FlowControllerV1.java
@@ -13,24 +13,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.taobao.csp.sentinel.dashboard.view;
+package com.taobao.csp.sentinel.dashboard.controller;
 
 import java.util.Date;
 import java.util.List;
 
 import com.alibaba.csp.sentinel.util.StringUtil;
 
+import com.taobao.csp.sentinel.dashboard.client.SentinelApiClient;
 import com.taobao.csp.sentinel.dashboard.datasource.entity.rule.FlowRuleEntity;
+import com.taobao.csp.sentinel.dashboard.discovery.MachineInfo;
+import com.taobao.csp.sentinel.dashboard.domain.Result;
 import com.taobao.csp.sentinel.dashboard.repository.rule.InMemoryRuleRepositoryAdapter;
-import com.taobao.csp.sentinel.dashboard.rule.DynamicRuleProvider;
-import com.taobao.csp.sentinel.dashboard.rule.DynamicRulePublisher;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -39,42 +38,38 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 /**
- * Flow rule controller (v2).
+ * Flow rule controller.
  *
+ * @author leyou
  * @author Eric Zhao
- * @since 1.4.0
  */
 @RestController
-@RequestMapping(value = "/v2/flow")
-public class FlowControllerV2 {
+@RequestMapping(value = "/v1/flow")
+public class FlowControllerV1 {
 
-    private final Logger logger = LoggerFactory.getLogger(FlowControllerV2.class);
+    private final Logger logger = LoggerFactory.getLogger(FlowControllerV1.class);
 
     @Autowired
     private InMemoryRuleRepositoryAdapter<FlowRuleEntity> repository;
 
     @Autowired
-    @Qualifier("flowRuleDefaultProvider")
-    private DynamicRuleProvider<List<FlowRuleEntity>> ruleProvider;
-    @Autowired
-    @Qualifier("flowRuleDefaultPublisher")
-    private DynamicRulePublisher<List<FlowRuleEntity>> rulePublisher;
+    private SentinelApiClient sentinelApiClient;
 
     @GetMapping("/rules")
-    public Result<List<FlowRuleEntity>> apiQueryMachineRules(@RequestParam String app) {
+    public Result<List<FlowRuleEntity>> apiQueryMachineRules(@RequestParam String app,
+                                                             @RequestParam String ip,
+                                                             @RequestParam Integer port) {
         if (StringUtil.isEmpty(app)) {
             return Result.ofFail(-1, "app can't be null or empty");
         }
+        if (StringUtil.isEmpty(ip)) {
+            return Result.ofFail(-1, "ip can't be null or empty");
+        }
+        if (port == null) {
+            return Result.ofFail(-1, "port can't be null");
+        }
         try {
-            List<FlowRuleEntity> rules = ruleProvider.getRules(app);
-            if (rules != null && !rules.isEmpty()) {
-                for (FlowRuleEntity entity : rules) {
-                    entity.setApp(app);
-                    if (entity.getClusterConfig() != null && entity.getClusterConfig().getFlowId() != null) {
-                        entity.setId(entity.getClusterConfig().getFlowId());
-                    }
-                }
-            }
+            List<FlowRuleEntity> rules = sentinelApiClient.fetchFlowRuleOfMachine(app, ip, port);
             rules = repository.saveAll(rules);
             return Result.ofSuccess(rules);
         } catch (Throwable throwable) {
@@ -84,11 +79,14 @@ public class FlowControllerV2 {
     }
 
     private <R> Result<R> checkEntityInternal(FlowRuleEntity entity) {
-        if (entity == null) {
-            return Result.ofFail(-1, "invalid body");
-        }
         if (StringUtil.isBlank(entity.getApp())) {
             return Result.ofFail(-1, "app can't be null or empty");
+        }
+        if (StringUtil.isBlank(entity.getIp())) {
+            return Result.ofFail(-1, "ip can't be null or empty");
+        }
+        if (entity.getPort() == null) {
+            return Result.ofFail(-1, "port can't be null");
         }
         if (StringUtil.isBlank(entity.getLimitApp())) {
             return Result.ofFail(-1, "limitApp can't be null or empty");
@@ -141,55 +139,97 @@ public class FlowControllerV2 {
         entity.setResource(entity.getResource().trim());
         try {
             entity = repository.save(entity);
-            publishRules(entity.getApp());
         } catch (Throwable throwable) {
             logger.error("Failed to add flow rule", throwable);
             return Result.ofThrowable(-1, throwable);
         }
+        if (!publishRules(entity.getApp(), entity.getIp(), entity.getPort())) {
+            logger.error("Publish flow rules failed after rule add");
+        }
         return Result.ofSuccess(entity);
     }
 
-    @PutMapping("/rule/{id}")
-    public Result<FlowRuleEntity> apiUpdateFlowRule(@PathVariable("id") Long id, @RequestBody FlowRuleEntity entity) {
-        if (id == null || id <= 0) {
-            return Result.ofFail(-1, "Invalid id");
+    @PutMapping("/save.json")
+    public Result<FlowRuleEntity> updateIfNotNull(Long id, String app,
+                              String limitApp, String resource, Integer grade,
+                              Double count, Integer strategy, String refResource,
+                              Integer controlBehavior, Integer warmUpPeriodSec, Integer maxQueueingTimeMs) {
+        if (id == null) {
+            return Result.ofFail(-1, "id can't be null");
         }
-        FlowRuleEntity oldEntity = repository.findById(id);
-        if (oldEntity == null) {
-            return Result.ofFail(-1, "id " + id + " does not exist");
-        }
+        FlowRuleEntity entity = repository.findById(id);
         if (entity == null) {
-            return Result.ofFail(-1, "invalid body");
+            return Result.ofFail(-1, "id " + id + " dose not exist");
         }
-        entity.setApp(oldEntity.getApp());
-        entity.setIp(oldEntity.getIp());
-        entity.setPort(oldEntity.getPort());
-        Result<FlowRuleEntity> checkResult = checkEntityInternal(entity);
-        if (checkResult != null) {
-            return checkResult;
+        if (StringUtil.isNotBlank(app)) {
+            entity.setApp(app.trim());
         }
-
-        entity.setId(id);
+        if (StringUtil.isNotBlank(limitApp)) {
+            entity.setLimitApp(limitApp.trim());
+        }
+        if (StringUtil.isNotBlank(resource)) {
+            entity.setResource(resource.trim());
+        }
+        if (grade != null) {
+            if (grade != 0 && grade != 1) {
+                return Result.ofFail(-1, "grade must be 0 or 1, but " + grade + " got");
+            }
+            entity.setGrade(grade);
+        }
+        if (count != null) {
+            entity.setCount(count);
+        }
+        if (strategy != null) {
+            if (strategy != 0 && strategy != 1 && strategy != 2) {
+                return Result.ofFail(-1, "strategy must be in [0, 1, 2], but " + strategy + " got");
+            }
+            entity.setStrategy(strategy);
+            if (strategy != 0) {
+                if (StringUtil.isBlank(refResource)) {
+                    return Result.ofFail(-1, "refResource can't be null or empty when strategy!=0");
+                }
+                entity.setRefResource(refResource.trim());
+            }
+        }
+        if (controlBehavior != null) {
+            if (controlBehavior != 0 && controlBehavior != 1 && controlBehavior != 2) {
+                return Result.ofFail(-1, "controlBehavior must be in [0, 1, 2], but " + controlBehavior + " got");
+            }
+            if (controlBehavior == 1 && warmUpPeriodSec == null) {
+                return Result.ofFail(-1, "warmUpPeriodSec can't be null when controlBehavior==1");
+            }
+            if (controlBehavior == 2 && maxQueueingTimeMs == null) {
+                return Result.ofFail(-1, "maxQueueingTimeMs can't be null when controlBehavior==2");
+            }
+            entity.setControlBehavior(controlBehavior);
+            if (warmUpPeriodSec != null) {
+                entity.setWarmUpPeriodSec(warmUpPeriodSec);
+            }
+            if (maxQueueingTimeMs != null) {
+                entity.setMaxQueueingTimeMs(maxQueueingTimeMs);
+            }
+        }
         Date date = new Date();
-        entity.setGmtCreate(oldEntity.getGmtCreate());
         entity.setGmtModified(date);
         try {
             entity = repository.save(entity);
             if (entity == null) {
                 return Result.ofFail(-1, "save entity fail");
             }
-            publishRules(oldEntity.getApp());
         } catch (Throwable throwable) {
-            logger.error("Failed to update flow rule", throwable);
+            logger.error("save error:", throwable);
             return Result.ofThrowable(-1, throwable);
+        }
+        if (!publishRules(entity.getApp(), entity.getIp(), entity.getPort())) {
+            logger.info("publish flow rules fail after rule update");
         }
         return Result.ofSuccess(entity);
     }
 
-    @DeleteMapping("/rule/{id}")
-    public Result<Long> apiDeleteRule(@PathVariable("id") Long id) {
-        if (id == null || id <= 0) {
-            return Result.ofFail(-1, "Invalid id");
+    @DeleteMapping("/delete.json")
+    public Result<Long> delete(Long id) {
+        if (id == null) {
+            return Result.ofFail(-1, "id can't be null");
         }
         FlowRuleEntity oldEntity = repository.findById(id);
         if (oldEntity == null) {
@@ -197,15 +237,17 @@ public class FlowControllerV2 {
         }
         try {
             repository.delete(id);
-            publishRules(oldEntity.getApp());
         } catch (Exception e) {
             return Result.ofFail(-1, e.getMessage());
+        }
+        if (!publishRules(oldEntity.getApp(), oldEntity.getIp(), oldEntity.getPort())) {
+            logger.info("publish flow rules fail after rule delete");
         }
         return Result.ofSuccess(id);
     }
 
-    private void publishRules(/*@NonNull*/ String app) throws Exception {
-        List<FlowRuleEntity> rules = repository.findAllByApp(app);
-        rulePublisher.publish(app, rules);
+    private boolean publishRules(String app, String ip, Integer port) {
+        List<FlowRuleEntity> rules = repository.findAllByMachine(MachineInfo.of(app, ip, port));
+        return sentinelApiClient.setFlowRuleOfMachine(app, ip, port, rules);
     }
 }
