+++ sentinel-dashboard/src/main/java/com/taobao/csp/sentinel/dashboard/controller/ResourceController.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.taobao.csp.sentinel.dashboard.view;
+package com.taobao.csp.sentinel.dashboard.controller;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -23,7 +23,8 @@ import com.alibaba.csp.sentinel.command.vo.NodeVo;
 
 import com.taobao.csp.sentinel.dashboard.domain.ResourceTreeNode;
 import com.taobao.csp.sentinel.dashboard.client.SentinelApiClient;
-import com.taobao.csp.sentinel.dashboard.view.vo.ResourceVo;
+import com.taobao.csp.sentinel.dashboard.domain.Result;
+import com.taobao.csp.sentinel.dashboard.domain.vo.ResourceVo;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
