+++ sentinel-dashboard/src/main/java/com/taobao/csp/sentinel/dashboard/domain/Result.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.taobao.csp.sentinel.dashboard.view;
+package com.taobao.csp.sentinel.dashboard.domain;
 
 /**
  * @author leyou
