+++ sentinel-dashboard/src/main/java/com/taobao/csp/sentinel/dashboard/domain/vo/MachineInfoVo.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.taobao.csp.sentinel.dashboard.view.vo;
+package com.taobao.csp.sentinel.dashboard.domain.vo;
 
 import java.util.ArrayList;
 import java.util.Date;
